Structure
Before adding the real "Code" I have gone ahead and added the process in
comments and functions.
Among this was adding a new project "SmartFileOAuthTest" which is a
simple console application in which is an example of the process a
client application would take to authenticate and use the SmartFile API
from.